+++ game/board.py
@@ -20,10 +20,10 @@ class Board:
     
     def validate_word(self, start_location_x, start_location_y, word, orientation):
         for i, letter in enumerate(word):
-            if orientation == 'H':
+            if orientation == 'Horizontal':
                 location_x = start_location_x
                 location_y = start_location_y + i
-            elif orientation == 'V':
+            elif orientation == 'Vertical':
                 location_x = start_location_x + i
                 location_y = start_location_y
 
@@ -32,13 +32,68 @@ class Board:
         return True
 
     def validate_word_inside_board(self, word, location, orientation):
-        row, col = location
+        location_x, location_y = location
         word_length = len(word)
 
-        if orientation == "H":
-            return col + word_length <= 15
-        elif orientation == "V":
-            return row + word_length <= 15
+        if orientation == "Horizontal":
+            return location_y + word_length <= 15
+        elif orientation == "Vertical":
+            return location_x + word_length <= 15
 
     def validate_word_out_of_board(self, word, location, orientation):
         return not self.validate_word_inside_board(word, location, orientation)
+
+
+    def validate_word_horizontal(self, word, location, orientation):
+        location_x, location_y = location
+        word_length = len(word)
+        found_letter = False
+
+        for i in range(word_length):
+            actual_tile = self.grid[location_x][location_y + i].letter
+            if actual_tile is not None and actual_tile.letter.lower() == word[i]:
+                found_letter = True
+
+        return found_letter and self.validate_word_inside_board(word, location, orientation)
+
+
+    def validate_word_vertical(self, word, location, orientation):
+        location_x, location_y = location
+        word_length = len(word)
+        found_letter = False
+        for i in range(word_length):
+            actual_tile = self.grid[location_x + i][location_y].letter
+            if actual_tile is not None:
+                if actual_tile.letter.lower() == word[i]:
+                    found_letter = True
+        return found_letter and self.validate_word_inside_board(word, location, orientation)
+
+
+    def is_empty(self):
+        if self.grid[7][7].letter is None:
+            return True
+        else:
+            return False
+
+    def word_in_the_center(self, word, location, orientation):
+        location_x, location_y = location
+        is_horizontal = orientation == "Horizontal"
+        is_vertical = orientation == "Vertical"
+
+        if is_horizontal and location_x == 7:
+            return self.validate_word_inside_board(word, location, orientation)
+
+        if is_vertical and location_y == 7:
+            return self.validate_word_inside_board(word, location, orientation)
+
+        return False
+
+
+    def validate_word_place_board(self, word, location, orientation):
+        if self.is_empty() is True:
+            return self.word_in_the_center(word, location, orientation)
+        else:
+            if orientation == "Horizontal":
+                return self.validate_word_horizontal(word, location, orientation)
+            else:
+                return self.validate_word_vertical(word, location, orientation)